refactor Sender
