+++ Sender.py
@@ -27,51 +27,3 @@ class Sender:
 
         Sender.sock = s
         Sender.address = address
-        listener = keyboard.Listener(on_press=Sender.on_press, on_release=Sender.on_release)
-        listener.start()  # start to listen on a separate thread
-
-        listener2 = mouse.Listener(
-            on_move=Sender.on_move,
-            on_click=Sender.on_click)
-        listener2.start()
-
-        listener.join()
-        print("start listen")
-
-    @staticmethod
-    def on_press(key):
-        if key == keyboard.Key.esc:
-            return False  # stop listener
-        try:
-            k = key.char  # single-char keys
-        except Exception as e:
-            print(e)
-            k = key.name  # other keys
-        k = f"press {k}"
-        print('Key pressed: ' + k)
-        Sender.sock.sendto(k.encode(), (Sender.address[4][0], Sender.myport))
-
-    @staticmethod
-    def on_release(key):
-        if key == keyboard.Key.esc:
-            return False  # stop listener
-        try:
-            k = key.char  # single-char keys
-        except Exception as e:
-            print(e)
-            k = key.name  # other keys
-        k = f"release {k}"
-        print('Key released: ' + k)
-        Sender.sock.sendto(k.encode(), (Sender.address[4][0], Sender.myport))
-
-    @staticmethod
-    def on_move(x, y):
-        print('Pointer moved to {0}'.format((x, y)))
-        k = f"moved {x},{y}"
-        Sender.sock.sendto(k.encode(), (Sender.address[4][0], Sender.myport))
-
-    @staticmethod
-    def on_click(x, y, button, pressed):
-        print('{0} at {1}'.format('Pressed' if pressed else 'Released', (x, y)))
-        k = f"clicked {button.name},{pressed},{x},{y}"
-        Sender.sock.sendto(k.encode(), (Sender.address[4][0], Sender.myport))
